Home.js Separation
