+++ server/route.js
@@ -2,7 +2,6 @@ const express = require('express');
 const router = express.Router();
 const controller = require('./controller');
 
-// 여기서 API 주소를 관리할 수 있다!
 router.get('/post/get', controller.post.get);
 
 router.post('/post/add', controller.post.add);
