+++ src/components/index.js
@@ -1,2 +1,3 @@
 export {default as LoginModal} from './loginModal';
 export {default as SignUpModal} from './signUpModal';
+export {default as TestContents} from './testContents';
