+++ src/pages/diary.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Diary = () => {
+
+    return (
+        <div>
+
+        </div>
+    );
+}
+export default Diary;
