+++ src/pages/index.js
@@ -1,2 +1,4 @@
 export {default as Login} from './login';
 export {default as Home} from './home';
+export {default as Diary} from './diary';
+export {default as Write} from './write';
