+++ src/pages/write.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Write = () => {
+
+    return (
+        <div>
+
+        </div>
+    );
+}
+export default Write;
